Reformat structure of network calls
The createSocket() function is now used to create the socket that data will be written to by the client. This returns a simple file descrptor if the socket creation is a success, otherwise it returns -1.

instead of haveing a sortfilethread, we now have a sendFileData thread that needs to be implemented. This will take in the socket file descriptor and read in the lines from the file, sending them through the socket.

Comments have been added starting at line 243 which decribe the location where implementation must occur.

Whatever is after the comment on line 280 still needs to be looked at as this code mostly relates to sorting. The majority of which is not needed.