Implement connection watch and more descriptions
The client should be waiting for the server to respond that it has finished with a file. This has been implemented in the watchConnection() function that will be used in a thread. Also more changes have been mad in terms of proper code format.

Additional comments have been added that exaplain locations and execution points.

We still must work out transmitting lines in the sendFileData() thread function as well as creating functionality to request a "dump" from the server.