Complete task for flat directory
As of this commit the lines are able to be read and printed properly based on project specifications for a flat directory of files. There is still segfaulting issues as a result of nested directories, most likely due to the threadholder[] of each directory thread not holdin the proper tid. Focus efforts on debugging the program in the directory search.

Besides this bug the server and client are communicating and we are succesful in other test cases. More rigporous testing is obviously needed but we are close to completion.