Merge Sam's client with most recent version.
This new client is actually able to accomplish the task fully set off in the assignment.

With proper documentation the project is essentially finished and this even completes the first extra credit which uses semaphores to create a client pool.

Old files are kept in the /Old directory and are still kept for reasons of continuity.